Added new item 'Speed Scroll'
A new item, 'Speed Scroll', has been added to the game assets. This included updating the language files for both English and French, adding a new JSON model, and revising the item's properties in the ItemInit class and SpeedScrollItem class. This Scroll has a common rarity and is represented by the 'written_scroll' texture.